% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scldataR_metadata.R
\name{get_countries_soda}
\alias{get_countries_soda}
\title{Get Countries}
\usage{
get_countries_soda()
}
\value{
A data frame containing country data.
}
\description{
This function retrieves country data from a CSV file stored in a GitHub repository.
The CSV file contains information related to countries.
}
\examples{
\dontrun{
country_data <- get_countries_soda()
}
}
