% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scldataR_map.R
\name{get_map}
\alias{get_map}
\title{Get geojson map.}
\usage{
get_map(level = "All", isoalpha3 = "All")
}
\arguments{
\item{level}{Optional.  (1, 2)}

\item{isoalpha3}{Optional. countries (alpha-3 country code)}
}
\value{
A data frame with selected indicators
}
\description{
Get geojson map.
}
\examples{
get_map(level='1',isoalpha3="ARG")
}
