% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scldataR.R
\name{query_dictionary}
\alias{query_dictionary}
\title{Get dictionary}
\usage{
query_dictionary(indicator = "All", collection = "All", resource = "All")
}
\arguments{
\item{indicator}{Selected indicator}

\item{collection}{Optional categories ()}

\item{resource}{Optional.}

\item{theme_en}{Optional.}

\item{theme_es}{Optional.}
}
\value{
A data frame with dictionary
}
\description{
Get dictionary
}
\examples{
query_dictionary()
query_dictionary(indicator="pobreza")
query_dictionary(collection="Household Socio-Economic Surveys")
}
