% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scldataR.R
\name{query_indicator}
\alias{query_indicator}
\title{Get available data about selected indicator.}
\usage{
query_indicator(
  indicator,
  categories = "All",
  countries = "All",
  yearstart = "All",
  yearend = "All",
  year = "All"
)
}
\arguments{
\item{indicator}{Selected indicator}

\item{categories}{Optional categories (sex, age, area)}

\item{countries}{Optional. countries (alpha-3 country code)}

\item{year}{Optional.}
}
\value{
A data frame with selected indicators
}
\description{
Get available data about selected indicator.
}
\examples{
query_indicator(indicator="pobreza,pindi_ci")
query_indicator(indicator="pobreza",categories='sex,age')
query_indicator(indicator="pobreza",categories='sex',countries='mex,arg')
}
