% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scldata_sodaR.R
\name{query_indicator_soda}
\alias{query_indicator_soda}
\title{Get available data about selected indicator from the IADB resource.}
\usage{
query_indicator_soda(
  indicator,
  categories = NULL,
  countries = "All",
  yearstart = "All",
  yearend = "All",
  latest = FALSE
)
}
\arguments{
\item{indicator}{Selected indicator}

\item{categories}{Optional categories (area, sex, quintile, education_level, age, ethnicity)}

\item{countries}{Optional. countries (alpha-3 country code)}

\item{yearstart}{Optional. Start of the year range}

\item{yearend}{Optional. End of the year range}

\item{latest}{Optional. Latest data point available}
}
\value{
A data frame with selected indicators
}
\description{
Get available data about selected indicator from the IADB resource.
}
\examples{
query_indicator_soda(indicator="pobreza", categories=c('sex','age'), countries='MEX')
query_indicator_soda(indicator="tasa_ocupacion", categories=c('sex','ethnicity'))
}
