% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scldata_dict}
\alias{scldata_dict}
\title{SCLData-Indicators dictionary.}
\format{
Dictionary
\describe{
\item{collection}{collection}
\item{resource}{resource}
\item{source}{source}
\item{theme_en}{theme_en}
\item{theme_es}{theme_es}
\item{indicator}{indicator}
\item{label_en}{label_en}
\item{label_es}{label_es}
\item{description_en}{description_en}
\item{description_es}{description_es}
\item{valuetype}{valuetype}
\item{categories}{categories}
}
}
\source{
\url{https://scldata.iadb.org/}
}
\usage{
scldata_dict
}
\description{
Dataset containing labels and descriptions of scldata indicators
}
\keyword{datasets}
