% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scldataR.R
\name{search_indicator}
\alias{search_indicator}
\title{Search for indicator using words and filters.}
\usage{
search_indicator(
  search,
  categories = "All",
  countries = "All",
  yearstart = "All",
  yearend = "All",
  year = "All"
)
}
\arguments{
\item{search}{search}

\item{countries}{Optional. countries (alpha-3 country code)}

\item{yearstart}{Optional. yearstart}

\item{yearend}{Optional. yearend}
}
\value{
A data frame with found results
}
\description{
Search for indicator using words and filters.
}
\examples{
search_indicator(search="pobreza")
}
